+++ packages/wallet/src/userop/client.ts
@@ -10,7 +10,6 @@ import {
 } from "userop";
 import { AATransactionReceipt } from '@0xsodium/transactions';
 import { SodiumJsonRpcProvider } from "./jsonrpc";
-import { TransactionReceipt } from '@ethersproject/abstract-provider';
 import { hexValue } from "ethers/lib/utils";
 
 export interface IUserOperation extends OriginIUserOperation {