+++ packages/multicall/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @0xsodium/multicall
 
+## 1.45.0
+
+### Minor Changes
+
+- fix sign message
+
+### Patch Changes
+
+- Updated dependencies
+  - @0xsodium/network@1.45.0
+  - @0xsodium/abi@1.45.0
+  - @0xsodium/utils@1.45.0
+
 ## 1.44.0
 
 ### Minor Changes